requirements.txt
fastapi
uvicorn

langgraph
langchain
langsmith
langchain-openai
langchain-community

psycopg2-binary
pgvector
sqlalchemy

pydantic
python-dotenv

tavily-python